V1.8 development

+++ erp-webapi/Models/EditModels/SprintReportEditModel.cs
@@ -31,7 +31,7 @@
         
         public int? BestCase { get; set; }
         
-        public int? Score { get; set; }
+        public decimal? Score { get; set; }
         
         public bool Essential { get; set; }
     }
